move ActiveRecordHelpers under Aggregate module
