+++ test/aggregate_test.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative './test_helper'
+
 class AggregateTest < ActiveSupport::TestCase
   test "truth" do
     assert_kind_of Module, Aggregate
