+++ test/unit/container_test.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
 require_relative '../test_helper'
-require 'test_after_commit' if Rails::VERSION::MAJOR === 4
-
+Aggregate::ActiveRecordHelpers::Version.if_version(
+  active_record_4: -> { require 'test_after_commit' }
+)
 class Aggregate::ContainerTest < ActiveSupport::TestCase
 
   class ActiveRecordStub
